refactor: Refactoring root component and breaking it down into smaller components/hooks
