+++ app/root.tsx
@@ -1,39 +1,21 @@
-import { StylesPlaceholder } from "@mantine/remix"
 import {
   Alert,
   Center,
+  Container,
   createEmotionCache,
-  Global,
-  IconBookOpen,
-  IconMail,
-  MantineProvider,
-  MantineThemeOverride,
-  theme,
 } from "@pokt-foundation/pocket-blocks"
 import { LinksFunction, LoaderFunction, MetaFunction, json } from "@remix-run/node"
-import {
-  Links,
-  LiveReload,
-  Meta,
-  Outlet,
-  Scripts,
-  ScrollRestoration,
-  useCatch,
-  useLoaderData,
-  useLocation,
-  useSearchParams,
-} from "@remix-run/react"
-import React, { useEffect, useMemo } from "react"
+import { Outlet, useCatch, useLoaderData } from "@remix-run/react"
+import React from "react"
 import { Auth0Profile } from "remix-auth-auth0"
-import analyticsInit from "./utils/analytics"
 import { authenticator } from "./utils/auth.server"
 import Footer, { links as FooterLinks } from "~/components/Footer"
 import Header, { links as HeaderLinks } from "~/components/Header"
-import { IconApp } from "~/components/Icons"
 import Nav, { links as NavLinks } from "~/components/Nav"
-import { FeatureFlagsContextProvider } from "~/context/FeatureFlagContext"
-import { TranslateContextProvider, useTranslate } from "~/context/TranslateContext"
-import { UserContextProvider } from "~/context/UserContext"
+import Document from "~/root/components/Document"
+import PlasmicContainer from "~/root/components/PlasmicContainer"
+import RootProviders from "~/root/components/RootProviders"
+import { useRoot } from "~/root/hooks/useRoot"
 import normalizeStyles from "~/styles/normalize.css"
 import rootStyles from "~/styles/root.css"
 import { getClientEnv } from "~/utils/environment.server"
@@ -80,199 +62,21 @@ export const loader: LoaderFunction = async ({ request }) => {
 
 createEmotionCache({ key: "pni" })
 
-export const portalTheme: MantineThemeOverride = {
-  ...theme,
-  primaryColor: "blue",
-  components: {
-    ...theme.components,
-    Paper: {
-      styles: {
-        root: {
-          overflow: "visible !important",
-        },
-      },
-    },
-    Card: {
-      styles: (theme) => ({
-        root: {
-          padding: "32px",
-          backgroundColor:
-            theme.colorScheme === "dark" ? theme.colors.navy[5] : theme.colors.gray[1],
-        },
-      }),
-    },
-    Tabs: {
-      styles: (theme) => ({
-        tabsList: {
-          borderBottom: "2px solid transparent",
-          marginBottom: theme.spacing.md,
-        },
-        tab: {
-          paddingRight: theme.spacing.xs,
-          paddingLeft: theme.spacing.xs,
-          transition: "border-color ease-in-out 0.3s, color ease-in-out 0.3s",
-          "&[data-active]": {
-            borderColor: theme.colors[theme.primaryColor][6],
-          },
-          "&:hover": {
-            backgroundColor: "transparent",
-            borderColor: "transparent",
-            color: theme.colors[theme.primaryColor][8],
-          },
-          "&[data-active]:hover": {
-            borderColor: theme.colors[theme.primaryColor][8],
-          },
-          "&:not(:last-child)": {
-            marginRight: theme.spacing.md,
-          },
-        },
-      }),
-    },
-    TextInput: {
-      styles: {
-        input: {
-          backgroundColor: "transparent",
-        },
-      },
-    },
-    Textarea: {
-      styles: {
-        input: {
-          backgroundColor: "transparent",
-        },
-      },
-    },
-    MultiSelect: {
-      styles: {
-        input: {
-          backgroundColor: "transparent",
-        },
-      },
-    },
-  },
-}
-
-const WithProviders = ({ children }: { children: React.ReactNode }) => {
-  return (
-    <MantineProvider
-      withCSSVariables
-      withGlobalStyles
-      withNormalizeCSS
-      theme={portalTheme}
-    >
-      <FeatureFlagsContextProvider>
-        <UserContextProvider>
-          <TranslateContextProvider>{children}</TranslateContextProvider>
-        </UserContextProvider>
-      </FeatureFlagsContextProvider>
-    </MantineProvider>
-  )
-}
-
-const Document = ({ children, title }: { children: React.ReactNode; title?: string }) => {
-  const { language } = useTranslate()
-  const [params] = useSearchParams()
-
-  useEffect(() => {
-    const referral = params.get("rid")
-    if (referral) {
-      window.localStorage.setItem("rid", referral)
-    }
-  }, [params])
-
-  return (
-    <html lang={language}>
-      <head>
-        <StylesPlaceholder />
-        <Global
-          styles={(theme) => ({
-            body: {
-              ...theme.fn.fontStyles(),
-              backgroundColor:
-                theme.colorScheme === "dark"
-                  ? theme.colors.navy[7]
-                  : theme.colors.gray[2],
-              color:
-                theme.colorScheme === "dark"
-                  ? theme.colors.gray[0]
-                  : theme.colors.navy[9],
-              lineHeight: theme.lineHeight,
-            },
-          })}
-        />
-        <Meta />
-        <Links />
-      </head>
-      <body>
-        {children}
-        <ScrollRestoration />
-        <Scripts />
-        <LiveReload />
-      </body>
-    </html>
-  )
-}
-
 export default function App() {
   const { ENV, user } = useLoaderData<RootLoaderData>()
-  const { t } = useTranslate()
-  const { pathname } = useLocation()
-  const isDashboard = useMemo(() => pathname.includes("/dashboard/"), [pathname])
-  useEffect(() => {
-    analyticsInit({ id: user?.id ?? "" })
-  }, [user])
-
-  const routes = useMemo(() => {
-    enum Protected {
-      Public = 0,
-      Private = 1,
-      PrivateAdmin = 2,
-      Admin = 3,
-    }
-
-    const allRoutes = [
-      {
-        to: "/dashboard/apps",
-        label: t.dashboard.routes.apps,
-        icon: IconApp,
-        protected: Protected.Public, // show this link to all. dashboard layout handles redirect to login.
-      },
-      {
-        to: "https://docs.pokt.network/",
-        external: true,
-        label: t.dashboard.routes.docs,
-        icon: IconBookOpen,
-        protected: Protected.Public,
-      },
-      {
-        to: "https://support.pokt.network",
-        external: true,
-        label: "Support",
-        icon: IconMail,
-        protected: Protected.Public,
-      },
-    ]
-
-    let protectedLevel = Protected.Public
-
-    if (user) {
-      protectedLevel = Protected.Private
-    }
-
-    return allRoutes.filter((r) => r.protected <= protectedLevel)
-  }, [t, user])
-
+  const { isDashboard, routes } = useRoot({ user })
   return (
     <>
       {isDashboard ? (
-        <WithProviders>
+        <RootProviders>
           <Document>
             <Header user={user}>
               <Nav ariaLabel="Main" routes={routes} />
             </Header>
             <main>
-              <Outlet />
-
+              <Container fluid className="container">
+                <Outlet />
+              </Container>
             </main>
             <Footer />
             <script
@@ -281,26 +85,9 @@ export default function App() {
               }}
             />
           </Document>
-        </WithProviders>
+        </RootProviders>
       ) : (
-        <html>
-          <head>
-            <StylesPlaceholder />
-            <Meta />
-            <Links />
-            <script
-              dangerouslySetInnerHTML={{
-                __html: `window.ENV = ${JSON.stringify(ENV)};`,
-              }}
-            />
-          </head>
-          <body>
-            <Outlet />
-            <ScrollRestoration />
-            <Scripts />
-            <LiveReload />
-          </body>
-        </html>
+        <PlasmicContainer />
       )}
     </>
   )
@@ -311,7 +98,7 @@ export const CatchBoundary = () => {
 
   if (caught.status === 404) {
     return (
-      <WithProviders>
+      <RootProviders>
         <Document title={`${caught.status} ${caught.statusText}`}>
           <Center className="error-container" mt="xl">
             <Alert color="red" title={`Application Error: ${caught.status}`}>
@@ -319,7 +106,7 @@ export const CatchBoundary = () => {
             </Alert>
           </Center>
         </Document>
-      </WithProviders>
+      </RootProviders>
     )
   }
   throw new Error(`Unexpected caught response with status: ${caught.status}`)
@@ -327,7 +114,7 @@ export const CatchBoundary = () => {
 
 export const ErrorBoundary = ({ error }: { error: Error }) => {
   return (
-    <WithProviders>
+    <RootProviders>
       <Document title="Uh-oh!">
         <div className="error-container">
           <dialog color="red" title="Application Error">
@@ -335,6 +122,6 @@ export const ErrorBoundary = ({ error }: { error: Error }) => {
           </dialog>
         </div>
       </Document>
-    </WithProviders>
+    </RootProviders>
   )
 }
