+++ app/root/components/Document/index.ts
@@ -1,0 +1,3 @@
+import Document from "./Document"
+export * from "./Document"
+export default Document
