+++ app/root/components/PlasmicContainer/PlasmicContainer.tsx
@@ -1,0 +1,35 @@
+import { StylesPlaceholder } from "@mantine/remix"
+import {
+  Links,
+  LiveReload,
+  Meta,
+  Outlet,
+  Scripts,
+  ScrollRestoration,
+} from "@remix-run/react"
+import React from "react"
+
+const PlasmicContainer = () => {
+  return (
+    <html>
+      <head>
+        <StylesPlaceholder />
+        <Meta />
+        <Links />
+        <script
+          dangerouslySetInnerHTML={{
+            __html: `window.ENV = ${JSON.stringify(ENV)};`,
+          }}
+        />
+      </head>
+      <body>
+        <Outlet />
+        <ScrollRestoration />
+        <Scripts />
+        <LiveReload />
+      </body>
+    </html>
+  )
+}
+
+export default PlasmicContainer
