+++ app/root/components/PlasmicContainer/index.ts
@@ -1,0 +1,3 @@
+import PlasmicContainer from "./PlasmicContainer"
+export * from "./PlasmicContainer"
+export default PlasmicContainer
