+++ app/root/components/RootProviders/index.ts
@@ -1,0 +1,3 @@
+import RootProviders from "./RootProviders"
+export * from "./RootProviders"
+export default RootProviders
